XIVY-19018 Removed StateOfActivity added AppState
