--- engine-cockpit/src/ch/ivyteam/enginecockpit/application/model/StateOfActivity.java
@@ -1,145 +1,0 @@
-package ch.ivyteam.enginecockpit.application.model;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.apache.commons.lang3.StringUtils;
-
-import ch.ivyteam.ivy.application.ActivityOperationState;
-import ch.ivyteam.ivy.application.ActivityState;
-import ch.ivyteam.ivy.application.IApplication;
-import ch.ivyteam.ivy.application.ReleaseState;
-
-public class StateOfActivity {
-  private String activityState;
-  private String activityStateIcon;
-  private String errorMessage;
-  private List<String> childProblems;
-  private String operation;
-  private String operationIcon;
-  private String releaseState = "";
-  private String releaseStateIcon;
-  private boolean processing;
-
-  public StateOfActivity() {
-    updateState(null);
-    updateOperation(null);
-  }
-
-  public StateOfActivity(IApplication activity) {
-    updateState(activity.getActivityState());
-    updateOperation(activity.getActivityOperationState());
-    this.errorMessage = activity.getErrorMessage();
-    this.childProblems = List.of();
-  }
-
-  public String getState() {
-    return activityState;
-  }
-
-  public String getStateCssClass() {
-    return activityState.toLowerCase();
-  }
-
-  public String getStateIcon() {
-    return activityStateIcon;
-  }
-
-  public String getOperation() {
-    return operation;
-  }
-
-  public String getOperationCssClass() {
-    return operation.toLowerCase();
-  }
-
-  public String getOperationIcon() {
-    return operationIcon;
-  }
-
-  public String getReleaseState() {
-    return releaseState;
-  }
-
-  public String getReleaseStateCssClass() {
-    return releaseState.toLowerCase();
-  }
-
-  public String getReleaseStateIcon() {
-    return releaseStateIcon;
-  }
-
-  public String getErrorMessage() {
-    return StringUtils.isBlank(errorMessage) ? "" : "\n" + errorMessage;
-  }
-
-  public String getChildProblems() {
-    var message = childProblems.stream().collect(Collectors.joining("\n"));
-    return StringUtils.isBlank(message) ? "" : "\n" + message;
-  }
-
-  public boolean isProcessing() {
-    return processing;
-  }
-
-  private void updateState(ActivityState update) {
-    if (update == null) {
-      update = ActivityState.INACTIVE;
-    }
-    this.activityState = update.name();
-    switch (update) {
-      case ACTIVE -> this.activityStateIcon = "ti ti-circle-check";
-      default -> this.activityStateIcon = "ti ti-player-pause";
-    }
-    ;
-  }
-
-  private void updateOperation(ActivityOperationState update) {
-    if (update == null) {
-      update = ActivityOperationState.INACTIVE;
-    }
-    this.operation = update.name();
-    this.processing = false;
-    switch (update) {
-      case ACTIVE -> this.operationIcon = "ti ti-circle-check";
-      case INACTIVE -> this.operationIcon = "ti ti-player-pause";
-      case ERROR -> this.operationIcon = "ti ti-circle-minus";
-      default -> {
-        this.operationIcon = "ti ti-refresh spinning";
-        this.processing = true;
-      }
-    }
-    ;
-  }
-
-  public void updateReleaseState(ReleaseState update) {
-    if (update == null) {
-      update = ReleaseState.DELETED;
-    }
-    this.releaseState = update.name();
-    switch (update) {
-      case RELEASED -> this.releaseStateIcon = "ti ti-circle-check";
-      case DEPRECATED -> this.releaseStateIcon = "ti ti-circle-half-vertical";
-      case ARCHIVED -> this.releaseStateIcon = "ti ti-archive";
-      case CREATED, PREPARED -> this.releaseStateIcon = "ti ti-speakerphone";
-      default -> this.releaseStateIcon = "ti ti-help-circle";
-    }
-    ;
-  }
-
-  public boolean is(ActivityState... states) {
-    return Arrays.asList(states).stream()
-        .anyMatch(s -> s.name().equals(activityState));
-  }
-
-  public boolean is(ReleaseState... states) {
-    return Arrays.asList(states).stream()
-        .anyMatch(s -> s.name().equals(releaseState));
-  }
-
-  public static boolean is(String operation, ActivityOperationState... states) {
-    return Arrays.asList(states).stream()
-        .anyMatch(s -> s.name().equals(operation));
-  }
-}
